feat(zsh): add omz plugins and improve installer to source ~/.zshrc at end

+++ zsh/install.zsh
@@ -1,15 +1,21 @@
+#!/bin/zsh
 if [ -f "$HOME/.zshrc" ]; then
-  echo "$PC[$SC*$PC] Found .zshrc, creating backup..."
+  echo "${PC}[$SC*$PC] Found .zshrc, creating backup..."
   cp $HOME/.zshrc $HOME/.zshrc.bk
 fi
 
-rm -r $HOME/.zshrc $HOME/aliases
+if [[ -f $HOME/.zshrc && -d $HOME/aliases ]]; then
+    rm -r $HOME/.zshrc $HOME/aliases
+fi
 
-echo "$PC[$SC*$PC] Copying zsh config..."
+echo "${PC}[$SC*$PC] Copying zsh config..."
 if [[ "$(pwd)" == *zsh ]]; then
   cp ./.zshrc $HOME/.zshrc
-  cp -r ./* $HOME
+  rsync -avq ./* $HOME --exclude .zshrc
 else
   cp ./zsh/.zshrc $HOME/.zshrc
-  cp -r ./zsh/* $HOME
+  rsync -avq ./zsh/* $HOME --exclude ./zsh/.zshrc
 fi
+
+echo "${PC}[$SC*$PC] Sourcing zsh config"
+source ~/.zshrc
